dom/webgpu/mochitest/test_insecure_context.html

<!-- This is somewhat redundant with
     dom/tests/mochitest/general/test_interfaces.js, but I think it's good to
     have something here as well. -->

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
  </head>
  <body>
    <script>
      ok(
        SpecialPowers.getBoolPref("dom.webgpu.enabled"),
        "Pref should be enabled."
      );
      ok(!isSecureContext, "test should not run in a secure context");
      ok(navigator.gpu === undefined, "navigator.gpu === undefined");
    </script>
  </body>
</html>
